Add ability to pass in arguments to resolve method
If user passes in an argument to the resolve method the library
shouldn't treat this as an error. This may not be the best practice but
just in case someone needs to wait till that promise has been resolved
to fire an action from there, this update will handle it
appropriately.