Allow Fields to have an "_error" assigned to them.
Fields can have object values, in which case we need a way to set an
error at the object level. The library will pass "_error" to the Field as
the "error" prop to be displayed, but isValid wasn't respecting it as an error.

Also reused `isValid` selector in the `reduxForm` `connect` function so
that we're not duplicating logic.

Fixes #1107