Improve dirty handling for check-boxes
`dirty(undefined, false)` should be `false`. It was `true`.

Intuitively, clicking twice on a check-box should return it to
pristine, which is false atm (unless the checkbox is initialized with
`false`).

Same behavior as empty and undefined for a string.

Also decreases code complexity. If both value and initial are falsey,
return “equal true”.

Comes with tests, both immutable and plain structures are covered.

I’d call it a safe merge and would be very pleased to see it shipped.
:hatching_chick: