fix checkbox behavior
Follow up to #1762
When a checkbox is initialized with undefined or is not initialized,
`false` should be treated as `undefined` and removed from the value
tree.
It follows the convention of text fields in redux-form (‘’ value is
removed when initial is `undefined`) and follows web form checkbox
original behavior (a checkbox is checked or ‘absent’).
Of course it won’t affect forms where checkboxes are initialized with
`false` (double-clicking them will bring back `false` instead of
`undefined`, on par with text fields in redux-form)