Ensure immutable state is merged into field meta props as plain object (#1877, #1955)
cb3abb5b41976458fc0c54246df2eeea1f538d34 added functionality that merged the state
into the meta object for fields. However, when using Immutable JS structures this
will be an Immutable.Iterable. This causes an exception in environments where
the Object.assign polyfill does not allow this (for instance in React Native)