Provide validation functions with access to form props
Currently, the API for `validate` passed to `config` allows props to be
accessible by validation rules. The same should be done for field-level
validations. This PR provides props passed to the form as the second argument
of the function returned by `generateValidator()`, making the APIs between the
two forms of validation consistent.