Fix initial submitError state for FieldArrays
When adding a new field to a FieldArray (which already has a field with a submitError), the submitError for the new field is initialised with `empty`.

This breaks form validation, as the new field has not been submitted yet and therefore should not have an initial submitError.

Therefore changing the initial value for `submitError` and `asyncError` of new fields to `undefined`.