add anyActive prop to reduxForm
Adding new prop passing to form via reduxForm. The value shows up when any field is active, but it is not passing to form at all. The prop, for example, helps doing more ux stuff like showing form as a modal when user is interacting withing the form and so on. Hope this proposal is worthy for you as well.