index.js.flow: Export FormProps we pass *to* components
Documentation states that 'FormProps' is what's passed *to* a component
wrapped with reduxForm(); however, what was exported here was the props
that should be passed *by* a user *to* the <Form> component.

Switched to using the same naming convention as is already used for
other prop-types with the same potential for confusion: exported
'FieldProps' for the props Field passes to user components, and
non-exported FieldInputProps for the props Field expects to get from its
callers.

Fixes #3823.