Use accurate Event type
The existing Event type would not typecheck for any correct
event-handling code; dataTransfer is not present for non-DnD events, and
value is not guaranteed to be present for non-Input events.

This change is quite safe:

1. Props the user is expected to pass in are replaced with supertypes of
what they were before (a function taking a subtype is a supertype)

2. Props the user is given are replaced with supertypes of what they
were before, but they should only get a type error in cases where they
are legitimately messing up (i.e. are passing an Event that doesn't have
an HTMLInputElement as its target to the handlers in FieldProps).