Re-throw errors if onSubmitFail is not provided
handleSubmit always returns/resolves with an error instead of re-throwing it. This seems to be a mistake, going by the comment above. Instead, return an appropriate error if there is an onSubmitFail callback and re-throw it otherwise.

This is probably a breaking change, as it can cause exceptions to be thrown on submission errors where they were not previously.